+++ src/core/Detecter.ts
@@ -2,10 +2,34 @@ import * as vscode from "vscode";
 
 export class Detecter {
 
-    private static methodPattern = /([\w_]+\([\w\s,:"=]*\))\s*\{/;
+    private static documentMethodMap = {}
+    private static methodPattern = /(([\w_]+)\s*\([\w\s,:"=]*\))\s*\{/;
     private static keywordPattern = /\b(if|While)\b/ig;
+
     /**
-     * detect method symbol by line
+     * detect method list by document
+     * @param document 
+     */
+    static getMethodList(document: vscode.TextDocument, usingCache = false): Method[] {
+
+        // if (usingCache && null != this.documentMethodMap[document.uri.path]) {
+        //     return this.documentMethodMap[document.uri.path];
+        // }
+
+        let methodList: Method[] = [];
+        const lineCount = Math.min(document.lineCount, 10000);
+        for (let line = 0; line < lineCount; line++) {
+            var method = Detecter.getMethodByLine(document, line)
+            if (method) {
+                methodList.push(method)
+            }
+        }
+        // this.documentMethodMap[document.uri.path] = methodList
+        return methodList;
+    }
+
+    /**
+     * detect method by line
      * @param document 
      * @param line 
      */
@@ -15,7 +39,7 @@ export class Detecter {
         const methodMatch = text.match(this.methodPattern);
         const keywordMatch = text.match(this.keywordPattern)
         if (methodMatch && !keywordMatch) {
-            return new Method(methodMatch[1], Detecter.getRemarkByLine(document, line - 1))
+            return new Method(methodMatch[1],methodMatch[2], line, Detecter.getRemarkByLine(document, line - 1))
         }
     }
 
@@ -34,5 +58,9 @@ export class Detecter {
 }
 
 export class Method {
-    constructor(public name: string, public comnent: string) { }
+    constructor(public full:string,public name: string, public line: number, public comnent: string) { }
+}
+
+export class FileChangeProvider{
+
 }