+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 2.0.3 (2020/3/10)
 - Fix if and while block as methods
 - Support complection method symbol.
+- Support go to method definition in same file.
 
 ## 2.0.2 (2019/11/27)
 - Enhance method symbol detection.